refactores posts to project posts
