+++ frontend/src/app/(appLayout)/projects/[projectId]/posts/[projectPostId]/page.tsx
