+++ frontend/src/components/project/PostForm.tsx
@@ -15,9 +15,12 @@ import { Textarea } from "../ui/textarea";
 import type { ProjectId } from "@/types/Project";
 
 function PostForm({ projectId }: { projectId: ProjectId }) {
-  const [formState, action] = useFormState(createProjectPost.bind(null, projectId), {
-    errors: {},
-  });
+  const [formState, action] = useFormState(
+    createProjectPost.bind(null, projectId),
+    {
+      errors: {},
+    },
+  );
   const formRef = useRef<HTMLFormElement>(null);
 
   return (
@@ -25,7 +28,9 @@ function PostForm({ projectId }: { projectId: ProjectId }) {
       ref={formRef}
       className="mt-2"
       action={async (formData) => {
-        formRef.current?.reset();
+        if (formState.success) {
+          formRef.current?.reset();
+        }
         action(formData);
       }}
     >
